Refactor some stuff into internal package
